[feat] Retriever 필터 메서드 안정화
