+++ src/main/java/org/project/domain/ai/rag/E/retrieve/file/DefaultMemoFileRetriever.java
@@ -1,6 +1,7 @@
-package org.project.domain.ai.rag.E.retrieve;
+package org.project.domain.ai.rag.E.retrieve.file;
 
 import lombok.RequiredArgsConstructor;
+import org.project.domain.ai.rag.A.extract.RagDocumentType;
 import org.project.domain.ai.rag.D.query.dto.RagQuery;
 import org.springframework.ai.document.Document;
 import org.springframework.ai.vectorstore.SearchRequest;
@@ -13,45 +14,44 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class DefaultMemoImageRetriever implements MemoImageRetriever {
+public class DefaultMemoFileRetriever implements MemoFileRetriever {
 
     private final VectorStore vectorStore;
 
     @Override
     public List<Document> retrieve(RagQuery query) {
 
-        SearchRequest searchRequest = SearchRequest.builder()
+        SearchRequest request = SearchRequest.builder()
                 .query(query.userPrompt())
-                .topK(3)
-                .similarityThreshold(0.2)
+                .topK(5)
                 .filterExpression(buildFilter(query))
                 .build();
 
-        return vectorStore.similaritySearch(searchRequest);
+        return vectorStore.similaritySearch(request);
     }
 
     private Filter.Expression buildFilter(RagQuery query) {
+
         FilterExpressionBuilder b = new FilterExpressionBuilder();
 
-        // 1️⃣ 기본 조건
         FilterExpressionBuilder.Op baseOp = b.and(
-                b.eq("type", "MEMO_IMAGE"),
+                b.eq("type", RagDocumentType.MEMO_FILE.name()),
                 b.eq("userId", query.userId())
         );
 
-        // 2️⃣ memoIds IN 조건
         if (query.memoIds() != null && !query.memoIds().isEmpty()) {
-
-            FilterExpressionBuilder.Op memoIdOp =
-                    b.in("memoId", query.memoIds().stream()
-                            .map(id -> (Object) id)
-                            .toList()
-                    );
-
-            return b.and(baseOp, memoIdOp).build();
+            return b.and(
+                    baseOp,
+                    b.in(
+                            "memoId",
+                            query.memoIds().stream()
+                                    .map(id -> (Object) id)
+                                    .toList()
+                    )
+            ).build();
         }
 
-        // 3️⃣ build() 필수
         return baseOp.build();
     }
+
 }
