+++ src/main/java/org/project/domain/ai/rag/E/retrieve/image/DefaultMemoImageRetriever.java
@@ -1,6 +1,7 @@
-package org.project.domain.ai.rag.E.retrieve;
+package org.project.domain.ai.rag.E.retrieve.image;
 
 import lombok.RequiredArgsConstructor;
+import org.project.domain.ai.rag.A.extract.RagDocumentType;
 import org.project.domain.ai.rag.D.query.dto.RagQuery;
 import org.springframework.ai.document.Document;
 import org.springframework.ai.vectorstore.SearchRequest;
@@ -13,42 +14,39 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class DefaultMemoFileRetriever implements MemoFileRetriever {
+public class DefaultMemoImageRetriever implements MemoImageRetriever {
 
     private final VectorStore vectorStore;
 
     @Override
     public List<Document> retrieve(RagQuery query) {
 
-        SearchRequest request = SearchRequest.builder()
+        SearchRequest searchRequest = SearchRequest.builder()
                 .query(query.userPrompt())
-                .topK(5)
+                .topK(3)
+                .similarityThreshold(0.2)
                 .filterExpression(buildFilter(query))
                 .build();
 
-        return vectorStore.similaritySearch(request);
+        return vectorStore.similaritySearch(searchRequest);
     }
 
     private Filter.Expression buildFilter(RagQuery query) {
 
         FilterExpressionBuilder b = new FilterExpressionBuilder();
 
-        // 1️⃣ 기본 조건
         FilterExpressionBuilder.Op baseOp = b.and(
-                b.eq("type", "MEMO_FILE"),
+                b.eq("type", RagDocumentType.MEMO_IMAGE.name()),
                 b.eq("userId", query.userId())
         );
 
-        // 2️⃣ memoIds IN 조건 (선택)
         if (query.memoIds() != null && !query.memoIds().isEmpty()) {
-
-            FilterExpressionBuilder.Op memoIdOp =
+            return b.and(
+                    baseOp,
                     b.in("memoId", query.memoIds().stream()
                             .map(id -> (Object) id)
-                            .toList()
-                    );
-
-            return b.and(baseOp, memoIdOp).build();
+                            .toList())
+            ).build();
         }
 
         return baseOp.build();
